eews_backend/stream/requirements.txt
absl-py==1.4.0
astunparse==1.6.3
async-timeout==4.0.2
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
confluent-kafka==2.2.0
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
distlib==0.3.7
dnspython==2.4.1
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.41.1
gast==0.4.0
gdown==4.7.1
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.56.2
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
influxdb==5.3.1
influxdb-client==1.37.0
kafka-python==2.0.2
keras==2.11.0
keras-swa==0.1.7
kiwisolver==1.4.4
libclang==16.0.6
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
msgpack==1.0.5
numpy==1.25.1
oauthlib==3.2.2
obspy==1.4.0
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
pause==0.3
Pillow==10.0.0
platformdirs==3.9.1
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pymongo==4.4.1
pyparsing==3.0.9
PySocks==1.7.1
pyspark==3.4.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
reactivex==4.0.4
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.1
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.2
